requirements.txt
twilio~=6.35.1
python-dotenv~=0.10.3
requests~=2.24.0
beautifulsoup4~=4.8.1
Flask~=1.1.1
firebase-admin
lxml